fix(environment): resolve module resolution error after renaming to EnvironmentContext
