+++ src/libs/FraudProtection/GroupIBSdkBridge/index.ts
@@ -1,4 +1,5 @@
-import {getEnvironment, getOldDotEnvironmentURL} from '@libs/Environment/Environment';
+import {getOldDotEnvironmentURL} from '@libs/Environment/Environment';
+import getEnvironment from '@libs/Environment/getEnvironment';
 import Log from '@libs/Log';
 import CONST from '@src/CONST';
 import cidMap from './cidMap';
